+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/KlumFactory.java
@@ -77,6 +77,7 @@ public class KlumFactory<T> {
         protected final B asPublicBuilder(InternalKlumBuilder<T> builder) {
             return (B) builder;
         }
+
     }
 
     /**